+++ src/utils/parse-arguments-into-options.ts
@@ -74,7 +74,7 @@ export function parseArgumentsIntoOptions(rawArgs: Args): {
   const program = new Command();
 
   program
-    .name("create-hbar")
+    .name("create-scaffold-hbar")
     .description("Scaffold a Hedera dApp project")
     .version(packageJson.version, "-v, --version", "Print CLI version and exit")
     .argument("[project-name]", "Name / path of the project directory to create")